Refactor connect_spec_version_instrument to tracing module
